requirements.txt
streamlit==1.24.0
yfinance==0.2.28
pandas==2.0.3
numpy==1.24.3
plotly==5.15.0
scikit-learn==1.3.0
ta==0.10.2
requests==2.31.0
python-dotenv==1.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
django==4.2.3
whitenoise==6.5.0
streamlit-option-menu==0.3.6
